Fix problem where server would not wait long enough for client to post pages
Under high load, we might ask a client to download 100 pages. The server
only waited 2 minutes before assuming a timeout; the client had no time
limit. Now we 1) make the time limit proportional to the number of pages
requested and 2) give the client the same time limit that the server has.