Do not hibernate on network error (except when getting instructions)
Previously, we assumed all network errors were due to the client being
offline (ie. client running on a laptop). Now, a network error is more
likely due to a proxy error. For a proxy error, there is no reason not to
retry soon.

If there is an issue getting the instructions, the client will still
hibernate for 2 minutes.