Scrape contact info for each Craigslist page
To get the contact info for a listing, we need to make a separate request.
After a scraped page is sent to the server, the server looks for a link to
the page with the contact info and then puts it in the queue.

Sometimes (at random), when we scrape the contact page, we are hit a
captcha page; in this case the server will keep the contact page in the
queue for a client to try again later.

Scraping contact pages might cause IP addresses to be flagged or blocked
more often, which would be bad for the proxies with static IPs. This has
not been checked yet.