Client: sleep(1) if you hit a network error or "connection reset by peer"
The current behavior we see is, sometimes the proxy server is either
broken or thinks we are making too many parallel requests, so requests
fail. Then, the clients quickly drain all queue items and fail them all.

Instead, let's go slow and wait for the proxies to start working again.