Remove the "doRepeat" client option and make that the default
We can just make all clients repeat (ask for more instructions instead of
exiting when done). The only downside is that the client will be running
most of the time, ie. that would not be good if running on a user's laptop.
(That approach doesn't scale anyway.)

Making the clients always repeat makes the mental model of how clients
work more clear.